refactor: fix typos accross ipynb and md files, fix mkdocs.yml, remove unused imports in py files
