+++ omnidocs/tasks/table_extraction/extractors/__init__.py
@@ -7,7 +7,7 @@ This module provides various table extractor implementations.
 from .camelot import CamelotExtractor
 from .pdfplumber import PDFPlumberExtractor
 from .ppstructure import PPStructureExtractor
-from .table_transfomer import TableTransformerExtractor
+from .table_transformer import TableTransformerExtractor
 from .tableformer import TableFormerExtractor
 from .tabula import TabulaExtractor
 from .surya_table import SuryaTableExtractor
@@ -16,8 +16,8 @@ __all__ = [
     'CamelotExtractor',
     'PDFPlumberExtractor',
     'PPStructureExtractor',
-    'TableTransformerExtractor',
+    'SuryaTableExtractor',
     'TableFormerExtractor',
+    'TableTransformerExtractor',
     'TabulaExtractor',
-    'SuryaTableExtractor'
 ]
