+++ omnidocs/tasks/text_extraction/extractors/__init__.py
@@ -13,22 +13,8 @@ from .surya_text import SuryaTextExtractor
 from .pdfplumber import PdfplumberTextExtractor
 PDFPLUMBER_AVAILABLE = True
 
-# Create aliases for backward compatibility and easier naming
-PyMuPDFExtractor = PyMuPDFTextExtractor
-PyPDF2Extractor = PyPDF2TextExtractor
-PDFTextExtractor = PdftextTextExtractor
-DoclingExtractor = DoclingTextExtractor
-
-# Aliases for all extractors
-PDFPlumberTextExtractor = PdfplumberTextExtractor
-
 # All extractors available
 __all__ = [
-    'PyMuPDFExtractor',
-    'PyPDF2Extractor',
-    'PDFPlumberTextExtractor',
-    'PDFTextExtractor',
-    'DoclingExtractor',
     'PyMuPDFTextExtractor',
     'PyPDF2TextExtractor',
     'PdfplumberTextExtractor',
